feat(embedded): Restructured codebase to be much more modular, still need to test code
