+++ embedded/esp/main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "main.c" "receiver.c"
+idf_component_register(SRCS "main.c" "motor_driver.c" "drone_processor.c"
                     INCLUDE_DIRS ".")
