+++ embedded/esp/main/drone_processor.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void drone_task(void *param);
