+++ embedded/esp/main/main.c
@@ -14,66 +14,9 @@
 #include <rmw_microros/rmw_microros.h>
 #endif
 
-#include "receiver.h"
-
-#include <drone_data/drone_data/msg/robot_position.h>
-
-#define RCCHECK(fn) { rcl_ret_t temp_rc = fn; if((temp_rc != RCL_RET_OK)){printf("Failed status on line %d: %d. Aborting.\n",__LINE__,(int)temp_rc);vTaskDelete(NULL);}}
-
-void subscription_callback(const void *msgin) {
-    const drone_data__msg__RobotPosition *msg = (const drone_data__msg__RobotPosition *)msgin;
-    printf("Received: %d\n",  (int)msg->x);
-}
-
-static void basic_task(void *param) {
-    rcl_allocator_t allocator = rcl_get_default_allocator();
-	rclc_support_t support;
-
-	rcl_init_options_t init_options = rcl_get_zero_initialized_init_options();
-	RCCHECK(rcl_init_options_init(&init_options, allocator));
-
-#ifdef CONFIG_MICRO_ROS_ESP_XRCE_DDS_MIDDLEWARE
-	rmw_init_options_t* rmw_options = rcl_init_options_get_rmw_init_options(&init_options);
-
-	// Static Agent IP and port can be used instead of autodisvery.
-	RCCHECK(rmw_uros_options_set_udp_address(CONFIG_MICRO_ROS_AGENT_IP, CONFIG_MICRO_ROS_AGENT_PORT, rmw_options));
-#endif
-
-	// create init_options
-	RCCHECK(rclc_support_init_with_options(&support, 0, NULL, &init_options, &allocator));
-
-    // create node
-	rcl_node_t node;
-	RCCHECK(rclc_node_init_default(&node, "esp32_example_node", "", &support));
-
-    // create subscriber
-    rcl_subscription_t subscriber;
-    RCCHECK(rclc_subscription_init_default(
-		&subscriber,
-		&node,
-		ROSIDL_GET_MSG_TYPE_SUPPORT(drone_data, msg, RobotPosition),
-		"esp32_example_subscriber"
-    ));
-
-    // create message
-    drone_data__msg__RobotPosition msg;
-
-    // create executor with a single handle
-    rclc_executor_t executor;
-    RCCHECK(rclc_executor_init(&executor, &support.context, 1, &allocator));
-
-    // add subscriber to executor
-    RCCHECK(rclc_executor_add_subscription(
-        &executor, 
-        &subscriber, 
-        &msg,
-		&subscription_callback, 
-        ON_NEW_DATA
-    ));
-
-    // run the executor forever (continuously receive messages)
-    rclc_executor_spin(&executor);
-}
+#include "motor_driver.h"
+#include "drone_processor.h"
+#include "util.h"
 
 void app_main(void)
 {
@@ -81,6 +24,6 @@ void app_main(void)
     ESP_ERROR_CHECK(uros_network_interface_initialize());
 #endif
 
-    xTaskCreate(basic_task, "basic_task", 4*1024, NULL, 1, NULL);
-    // xTaskCreate(receiving_task, "receiving_task", 4*1024, NULL, 1, NULL);
+    xTaskCreate(drone_task, "drone_task", 4*1024, NULL, 1, NULL);
+    xTaskCreate(motor_task, "motor_task", 4*1024, NULL, 1, NULL);
 }
